Fix charset issue. Move Java I/O related methods to separate thread.
